GIF display with time info
The script extracting 2D slices from 4d time-series image of 3D image
tpts can now get the units of time and time-step from the command line
and overlay the info on the gif animation.

1. The script is also updated with argparse from the old use of
optparse.
2. TODO: Move the 3D slice extractor to the project directory itself and
check first for the paths of the binaries used in this script: ANTS and
convert of imagemagick.