Changes the way Lame parameters are provided
1. When using command line options to set Lame parameters one should now
use comma to separate individual values and give no space. No need of
the quotes to treat it as a single string too. This prevents possible
problem with parsing of strings inside quote that was seen with mpiexec
of petsc.

2. The compute matrix in Taras solver now uses constant viscosity
discretization even when mu image is given. This is because the variable
viscosity case matrix building code has not been updated for long and is
perhaps not correct.

enter the commit message for your changes. Lines starting