Adds possibility to choose divergence stencil
In PetscAdlemtaras3d, a bool variable to choose between two different
stencil for divergence operator in the continuity equation is
available. Using 12 point stencil will make the divergence compatible
with the ITK divergence computation. With this option, the prescribed
atrophy will exactly match the negative of divergence when computed from
ITK too. Until now, this was not the case since the staggered grid
velocity on faces when interpolated to cell centres and provided to ITK,
it computes divergence differently. This difference is more pronounced
when the input prescribed atrophy/div is not continuous and has jumps.

Setting this bool variable false results in the solver same as before.

Also, in CSF regions it seems it is not necessary to set the equation
div(u) + kp = 0. Instead, it seems we can leave the rows corresponding
to the CSF cells for continuity equation to zero. This will lead to
faster solver convergence without affecting the velocity in tissue while
having a very small impact on the velocity in CSF regions. I have
explained this in comments but have not used this yet to be consistent
with previous sets of experiments.