Fixes major bug in solver for multiple time-steps.
When the solver is run for multiple time-steps with changing brain mask
the matrix A in solving Ax=b can change with the mask. The rows of A
corresponding to CSF regions for the continuity equation div(u) + kp = 0
have non-zero k values. Now if these rows now represent tissue regions
due to the changing mask, I should set the values of k to zero. I was
not doing this! I was only checking if the row corresponds to CSF and
set non-zero k, elsewhere I just did not set anything thinking that when
the computeOperator function is called to build matrix, it has zero
values everywhere. But this is evidently not the case. And this was the
reason why after certain number of solves, the convergence was very slow
when I solved the system for multiple number of times. But when I used
the same mask that was used in say, nth solve, as input to the model
starting as a first time step it converged much faster.

TODO: I still need to figure out why sometimes ITK raises exception when
I redistribute atrophy from CSF regions to nearby tissue regions when
doing multiple solves and warping the atrophy map of previous time-step
to get the new atrophy map.