Fixes bug in redistributing atrophy from CSF
1. When simulating for multiple time-step, the input atrophy is warped
by the obtained velocity field to get a new atrophy map for the next
step. This warping results in creating atrophy values in some CSF
regions. I was redistributing these atrophy to nearby tissue voxels but
there was a bug resulting in "out of bounds" error from ITK neigborhood
iterator. I found the bug. This was when CSF voxels in the mask were in
the image borders. So the neigbhorhood of such border voxels contained
positions outside images and by default the boundary condition used by
ITK is Neumann condition. So if these neighbors had tissue voxel as
adjacent voxels, then these neighbors were considered as tissue voxels
too. Trying to set a new redistributed value to these voxels resulted in
the runtime exception since they are not valid regions for writing.

The solution I have now is to set the proper boundary condition for the
input mask. The outside border values are set to be constant and of
label CSF. This way redistributing atrophy will never try to set any
values on these positions since they have CSF labels and not tissue
labels.